requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
dnspython==2.3.0
emoji==0.5.4
mypy==1.1.1
mypy-extensions==1.0.0
pymongo==4.3.3
python-telegram-bot==13.14
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2